Move Namer to its own package to avoid cycle between packages
